+++ libs/localization/src/lib/en.json
@@ -1163,7 +1163,7 @@
       "cohortInfo": "Cohort info",
       "cohortList": "Cohort list",
       "deleteCohort": "Deleting {0}?",
-      "delteConfirm": "Are you sure you want to delete this cohort?",
+      "deleteConfirm": "Are you sure you want to delete this cohort?",
       "selectCohort": "Select a cohort",
       "shiftCohort": "Shift cohort",
       "shiftCohortDescription": "Select a cohort from the cohort list. Apply the cohort to the dashboard."
